+++ app/jobs/interrupt_llm_usages_job.rb
@@ -1,0 +1,7 @@
+class InterruptLlmUsagesJob < ApplicationJob
+  queue_as :default
+
+  def perform
+    LlmUsage.overdue.find_each(batch_size: 500, &:interrupt!)
+  end
+end